+++ include/entities/MeshGeometry_t.h
@@ -30,7 +30,7 @@ namespace APTracer { namespace Entities {
              */
             ~MeshGeometry_t();
 
-            unsigned int n_tris_; /**< @brief Number of triangular faces held by the mesh geometry.*/
+            size_t n_tris_; /**< @brief Number of triangular faces held by the mesh geometry.*/
             std::string* mat_; /**< @brief Array of strings representing each face's material's name. Size: n_tris_.*/
             Vec3f* v_; /**< @brief Array of points representing the triangular faces. Size: 3*n_tris_. Face i has the points v_[3*i], v_[3*i + 1], v_[3*i + 2].*/
             double* vt_; /**< @brief Array of uv coordinates representing the triangular faces' texture coordinates. Size: 6*n_tris_. Face i has the uvs [vt_[6*i], vt_[6*i+1]], [vt_[6*i+2], vt_[6*i+3]], [vt_[6*i+4], vt_[6*i+5]].*/
