+++ src/entities/MaterialMap_t.cpp
@@ -5,8 +5,8 @@
 using APTracer::Entities::MaterialMap_t;
 using APTracer::Entities::Material_t;
 
-MaterialMap_t::MaterialMap_t(std::string* names, Material_t** materials, unsigned int n) {
-    for (unsigned int i = 0; i < n; i++){
+MaterialMap_t::MaterialMap_t(std::string* names, Material_t** materials, size_t n) {
+    for (size_t i = 0; i < n; i++){
         materials_[names[i]] = materials[i];
     }
 }
